Create MPU6050IMU
Added Madgwick's open source quaternion filter of gyro and accelerometer data toproduce a quaternion estimate of relative orientation. The algorithm outputs yaw, pitch, and roll in the AHRS framework but suffers from yaw drift of about one degree per minute or less. This despite including the gyro drift filter and bias compensation. This is state of the art without a magnetic reference point to maintain absolute orientation. Still the algorithm does a pretty good job estimating yaw, which in principle, it should not be able to do well.